Replaced some more slow string indices by std::map. Fixed once again signed/unsigned issues
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/trunk@319 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ course_render.cpp
@@ -51,10 +51,10 @@ void RenderCourse () {
 //				DrawTrees
 // --------------------------------------------------------------------
 void DrawTrees() {
-    int			tree_type = -1;
-    int       	item_type = -1;
-	TObjectType	*object_types = &Course.ObjTypes[0];
-	CControl *ctrl = Players.GetCtrl (g_game.player_id);
+    size_t			tree_type = -1;
+    size_t			item_type = -1;
+	TObjectType*	object_types = &Course.ObjTypes[0];
+	CControl*		ctrl = Players.GetCtrl (g_game.player_id);
 
 	set_gl_options (TREES);
